warehouse/cloudquery-dune-contract-usage/pyproject.toml
[tool.poetry]
name = "dune-contract-usage"
version = "1.0.0"
description = "Collect contract usage from dune"
authors = ["Kariba Labs"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "dune_contract_usage" }]

[tool.poetry.dependencies]
python = "^3.11"
cloudquery-plugin-sdk = "^0.1.12"
arrow = "^1.3.0"

[tool.poetry.scripts]
dune_contract_usage = 'dune_contract_usage.serve:run'
dune_contract_usage_test = 'dune_contract_usage.load_csvs:run'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
